# optimize: implement unknown field function without serialization